split files and change Makefile
