+++ Makefile
@@ -1,6 +1,21 @@
-project1 : HelloGraphics.cpp
-	g++ -std=c++11 -o $@ $< -lGL -lGLU -lglut
+CC=g++
+CFLAGS=-c -std=c++11
+LIBS=-lGL -lGLU -lglut
+EXECUTABLE=project1
+SOURCES=main.cpp data.cpp graphics.cpp
+OBJECTS=$(SOURCES:.cpp=.o)
+
+# The following is completely generic
+
+all: $(SOURCES) $(EXECUTABLE)
+	@echo $(EXECUTABLE) has been compiled
+
+compile:
+	$(CC) $(CFLAGS) $< -o $@
+
+$(EXECUTABLE): $(OBJECTS) 
+	$(CC) $(LIBS) $(OBJECTS) -o $@ 
 
 clean:
-	rm project1 
-	rm *.o 
+	$(RM) *.o $(EXECUTABLE)
+	@echo $(EXECUTABLE) has been cleaned
